requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
PyJWT ==2.1.0
pytz==2021.1
sqlparse==0.4.2 
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==3.13.0
drf-writable-nested==0.7.0